+++ src/config.ts
@@ -107,6 +107,7 @@ const modeConfig = {
 	 */
 	tokens: {
 		apostrophes: 'mw-apostrophes',
+		apostrophesLink: 'mw-apostrophes-link',
 		comment: 'mw-comment',
 		doubleUnderscore: 'mw-double-underscore',
 		em: 'mw-em',